requirements.txt
python-dotenv
requests
docker
